fix: 라벨 파일 라우팅 구조 변경 (#46)
